requirements.txt
torch==1.12.1
Pillow==9.2.0
torchvision==0.13.1
scikit-learn==1.1.2
pycocotools==2.0.5
opencv-python==4.6.0.66
matplotlib==3.6.0



